requirements.txt
pytest==8.3.5
requests==2.32.0
allure-pytest==2.15.0
allure-python-commons==2.15.0